move out bootstrap token utils to k8s.io/cluster-bootstrap
