+++ cmd/kubeadm/app/cmd/token_test.go
@@ -31,6 +31,7 @@ import (
 	"k8s.io/client-go/kubernetes/fake"
 	core "k8s.io/client-go/testing"
 	"k8s.io/client-go/tools/clientcmd"
+	bootstraptokenapi "k8s.io/cluster-bootstrap/token/api"
 	kubeadmapiv1beta2 "k8s.io/kubernetes/cmd/kubeadm/app/apis/kubeadm/v1beta2"
 	outputapischeme "k8s.io/kubernetes/cmd/kubeadm/app/apis/output/scheme"
 	outputapiv1alpha1 "k8s.io/kubernetes/cmd/kubeadm/app/apis/output/v1alpha1"
@@ -156,7 +157,7 @@ func TestRunCreateToken(t *testing.T) {
 	}
 	for _, tc := range testCases {
 		t.Run(tc.name, func(t *testing.T) {
-			bts, err := kubeadmapiv1beta2.NewBootstrapTokenString(tc.token)
+			bts, err := bootstraptokenapi.NewBootstrapTokenString(tc.token)
 			if err != nil && len(tc.token) != 0 { // if tc.token is "" it's okay as it will be generated later at runtime
 				t.Fatalf("token couldn't be parsed for testing: %v", err)
 			}
@@ -427,7 +428,7 @@ abcdef.1234567890123456   <forever>   <never>   signing,authentication   valid b
 		t.Run(tc.name, func(t *testing.T) {
 			token := outputapiv1alpha1.BootstrapToken{
 				BootstrapToken: kubeadmapiv1beta2.BootstrapToken{
-					Token:       &kubeadmapiv1beta2.BootstrapTokenString{ID: tc.id, Secret: tc.secret},
+					Token:       &bootstraptokenapi.BootstrapTokenString{ID: tc.id, Secret: tc.secret},
 					Description: tc.description,
 					Usages:      tc.usages,
 					Groups:      tc.extraGroups,
