+++ staging/src/k8s.io/cluster-bootstrap/util/tokens/tokens.go
@@ -18,21 +18,23 @@ package tokens
 
 import (
 	"fmt"
-	"regexp"
 
 	"k8s.io/cluster-bootstrap/token/api"
-)
-
-var (
-	bootstrapTokenRe = regexp.MustCompile(api.BootstrapTokenPattern)
+	legacyutil "k8s.io/cluster-bootstrap/token/util"
 )
 
 // ParseToken tries and parse a valid token from a string.
 // A token ID and token secret are returned in case of success, an error otherwise.
 func ParseToken(s string) (tokenID, tokenSecret string, err error) {
-	split := bootstrapTokenRe.FindStringSubmatch(s)
+	split := api.BootstrapTokenRe.FindStringSubmatch(s)
 	if len(split) != 3 {
 		return "", "", fmt.Errorf("token [%q] was not of form [%q]", s, api.BootstrapTokenPattern)
 	}
 	return split[1], split[2], nil
+}
+
+// NewBootstrapTokenStringFromIDAndSecret is a wrapper around NewBootstrapTokenString
+// that allows the caller to specify the ID and Secret separately
+func NewBootstrapTokenStringFromIDAndSecret(id, secret string) (*api.BootstrapTokenString, error) {
+	return api.NewBootstrapTokenString(legacyutil.TokenFromIDAndSecret(id, secret))
 }
